backend/requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# OpenAI интеграция
openai==1.3.7

# Утилиты
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1

# Логирование
loguru==0.7.2

# CORS
fastapi-cors==0.0.6 